requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
branca==0.8.1
build==1.2.2.post1
buildozer==1.5.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.7
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
courlan==1.3.2
cryptography==44.0.2
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
dfdatetime==20240504
dfvfs==20240505
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
dtfabric==20230520
durationpy==0.9
faiss-cpu==1.10.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
folium==0.19.5
fpdf==1.7.2
frida==16.7.14
frida-tools==13.7.1
frozenlist==1.6.0
fsspec==2025.3.2
gguf==0.16.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.2
Kivy==2.3.1
Kivy-Garden==0.1.5
kubernetes==32.0.1
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.37
libbde-python==20240502
libcaes-python==20240413
libewf-python==20240506
libfcrypto-python==20240414
libfsapfs-python==20240429
libfsext-python==20240501
libfsfat-python==20240501
libfshfs-python==20240501
libfsntfs-python==20240501
libfsxfs-python==20240501
libfvde-python==20240502
libfwnt-python==20240415
libluksde-python==20240503
libmodi-python==20240507
libphdi-python==20240508
libqcow-python==20240308
libsigscan-python==20240505
libsmdev-python==20240505
libsmraw-python==20240506
libvhdi-python==20240509
libvmdk-python==20240510
libvsapm-python==20240503
libvsgpt-python==20240504
libvshadow-python==20240504
libvslvm-python==20240504
llama_cpp_python==0.3.8
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.32.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.21.1
openai==1.77.0
opencv-python==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.52.0
pluggy==1.5.0
posthog==4.0.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyNaCl==1.5.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pysqlite3==0.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-nmap==0.7.1
python-socketio==5.13.0
pytsk3==20250312
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
-e git+https://github.com/RafalW3bCraft/WhisperAiEngine.git@ef1f3eea5fb80686c9a5156569b37b4593d3e87e#egg=repl_nix_workspace
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==12.6.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.1.0
sh==2.2.2
shell-genie==0.2.10
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlmap==1.9.4
starlette==0.45.3
streamlit==1.44.0
streamlit_folium==0.24.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tld==0.13
tokenizers==0.21.1
toml==0.10.2
torch==2.5.0
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.51.3
typer==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xattr==1.1.4
xyzservices==2025.1.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
